Adds Nav Bar to webpage
