+++ src/components/App.js
@@ -1,0 +1,17 @@
+import logo from './logo.svg';
+import './App.css';
+import NavBar from './NavBar';
+
+function App() {
+  return (
+    <div className="App">
+       <NavBar/>
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        
+      </header>
+    </div>
+  );
+}
+
+export default App;
